Triage fixes: capture the payload, and keep Supabase out of the parser
Two defects found reviewing the branch, both of which would have shipped.

The capture stored no payload at all. It read `entry.jsonString`, which
the LogEntry type declares but the decoder never sets — the field it
emits is `text`. Every captured entry would have arrived with label,
hash, size and position and nothing to read, which is the one thing the
feature exists for. The type now documents which of the two is real, and
the same mistake turns out to predate this branch: the summarised-GRE
guard in GreToClient reads jsonString too, so it has never once fired.

The parser half no longer imports Supabase. Reaching isLabelWanted
through a module that also imported the client meant the background
window constructed a second auth client against the same stored session
— a token-refresh race, on the window that must never stall. The network
half is now `logCaptureSync`, imported only by the main window, and the
parser half imports nothing at all (which its test now demonstrates by
needing no mocks).
