+++ src/broadcastChannel/mainChannelListeners.ts
@@ -4,7 +4,7 @@ import { overlayTitleToId } from "../common/maps";
 import { LOGIN_OK } from "../constants";
 import { pushDraft, pushFormatsSnapshot } from "../data/cloudSync";
 import { isDraftDeleted } from "../data/deletedDrafts";
-import { submitLogCapture } from "../data/logCapture";
+import { submitLogCapture } from "../data/logCaptureSync";
 import setDbMatch from "../data/setDbMatch";
 import { getUserNamespacedKey, putData } from "../data/store";
 import syncDrafts from "../data/syncDrafts";
